views.py refactoring
